74-Policiticas de acceso a roles

+++ app/Http/Controllers/Admin/RolesController.php
@@ -12,6 +12,8 @@ class RolesController extends Controller
 {
 	public function index()
 	{
+		$this->authorize('view', new Role);
+		
 		return view('admin.roles.index', [
 			'roles' => Role::all()
 		]);
@@ -19,6 +21,8 @@ class RolesController extends Controller
 
 	public function create()
 	{
+		$this->authorize('create', $role = new Role);
+
 		return view('admin.roles.create', [
 			'role' => New Role,
 			'permissions' => Permission::pluck('name', 'id')
@@ -37,6 +41,8 @@ class RolesController extends Controller
 			]
 		); */
 
+		$this->authorize('create', new Role);
+
 		$role = Role::create($request->validated());
 
 		if ($request->has('permissions'))
@@ -47,13 +53,10 @@ class RolesController extends Controller
 		return redirect()->route('admin.roles.index')->withFlash('El role fue creado correctamente');
 	}
 
-    public function show($id)
-    {
-        //
-    }
-
 	public function edit(Role $role)
 	{
+		$this->authorize('update', $role);
+
 		return view('admin.roles.edit', [
 			'role' => $role,
 			'permissions' => Permission::pluck('name', 'id')
@@ -70,6 +73,8 @@ class RolesController extends Controller
 			]
 		); */
 
+		$this->authorize('update', $role);
+
 		$role->update($request->validated());
 
 		$role->permissions()->detach();		//Quitar los permisos
@@ -84,10 +89,7 @@ class RolesController extends Controller
 
 	public function destroy(Role $role)
 	{
-		if ($role->id === 1)
-		{
-			throw new \Illuminate\Auth\Access\AuthorizationException('No se puede eliminar este rol.');
-		};
+		$this->authorize('delete', $role);
 
 		$role->delete();
